Instagram.space.bot/requirements.txt
instabot
requests
python-dotenv
